// ui/src/interfaces/FootballInterface.tsx
export interface CountryInterface {
    id: string;
    name: string;
    slug: string;
    code?: string | null;
    dial_code?: string | null;
    flag?: string | null;
    logo?: string | null;
    continent_id: string;
    has_competitions: boolean;
    competitions: CompetitionInterface[];
    priority_number: number;
    status_id: string;
    user_id: string;
    created_at: string;
    updated_at: string;
}
// ...
export interface StandingInterface {
    id: string;
    competition_id: string;
    season_id: string;
    season: SeasonInterface;
    stage: string;
    type: string;
    group: string | null;
    status_id: string;
    user_id: string;
    created_at: string;
    updated_at: string;
    standing_table: StandingTableInterface[];
}

export interface SeasonInterface {
    id: string | number;
    competition_id: string;
    start_date: string;
    end_date: string;
    is_current: number;
    current_matchday: number | null;
    total_matchdays: number | null;
    winner_id: string | null;
    stages: any[]; // Define the structure based on your data
    status_id: string;
    user_id: string;
    created_at: string;
    updated_at: string;
}

export interface LastAction {
    id: number;
    competition_id: number;
    seasons_last_fetch: string | null;
    standings_recent_results_last_fetch: string | null;
    standings_historical_results_last_fetch: string | null;
    matches_recent_results_last_fetch: string | null;
    matches_historical_results_last_fetch: string | null;
    matches_fixtures_last_fetch: string | null;
    matches_shallow_fixtures_last_fetch: string | null;
    match_recent_results_last_fetch: string | null;
    match_historical_results_last_fetch: string | null;
    match_fixtures_last_fetch: string | null;
    match_shallow_fixtures_last_fetch: string | null;
    predictions_last_train: string | null;
    predictions_trained_to: string | null;
    predictions_last_done: string | null;
    stats_last_done: string | null;
    predictions_stats_last_done: string | null;
    created_at: string;
    updated_at: string;
    [key: string]: string | number | null | undefined;
}

export interface CompetitionInterface {
    id: string;
    name: string;
    slug: string;
    code: string;
    type: string;
    logo: string | null;
    plan: string | null;
    abbreviation: string | null;
    has_teams: boolean | null;
    continent_id: string;
    country_id: string | null;
    country: CountryInterface;
    priority_number: number;
    predictions_last_train: string | null
    Predictions_last_train: string | null
    predictions_trained_to: string | null
    last_updated: string;
    last_fetch: string | null;
    last_detailed_fetch: string | null;
    image: string | null;
    stage_id: string;
    status_id: string;
    user_id: string;
    created_at: string;
    updated_at: string;
    standings: StandingInterface[];
    seasons: SeasonInterface[];
    games_per_season: number;
    available_seasons: number;
    last_action: LastAction
}
// ...
export interface CompetitionTabInterface {
    record: CompetitionInterface | undefined;
    getRecord?: () => void;
    seasons?: SeasonInterface[] | null
    selectedSeason?: SeasonInterface | null
    setSelectedSeason?: React.Dispatch<React.SetStateAction<SeasonInterface | null>>;
    mainKey?: number
    setMainKey?: React.Dispatch<React.SetStateAction<number>>;
    useDate?: boolean;
    isDisabled?: boolean
    setUseDates?: any
}

export interface SeasonsListInterface {
    seasons?: SeasonInterface[] | null
    selectedSeason?: SeasonInterface | null
    handleSeasonChange?: (season: SeasonInterface | null) => void;

}